requirements.txt
black==23.10.1
opencv-python==4.9.0.80
ruff==0.1.3
gdown==4.7.1
numpy==1.26.1
matplotlib==3.8.0
pandas==2.2.2
Pillow==10.1.0
pydot==2.0.0
scipy==1.13.0
